Move cuboid class into array/ directory, which will be populated as we deatch other classes from MATLAB
