+++ tdms/include/simulation_manager/objects_from_infile.h
@@ -10,8 +10,8 @@
 #include "matrix.h"
 
 #include "arrays.h"
+#include "arrays/cuboid.h"
 #include "cell_coordinate.h"
-#include "cuboid.h"
 #include "field.h"
 #include "fieldsample.h"
 #include "globals.h"
